Push avant modif

+++ srcs/requirements/wordpress/conf/config.sh
@@ -1,18 +1,33 @@
-cd /var/www/wordpress
+#!/bin/bash
 
-wp core config	--dghost=$DB_HOST \
-		--dbuser=$DB_USER \
-		--dbpass=$DB_PASSWORD \
-		--allow-root
+sleep 10
 
-wp core install	--title=$WP_TITLE \
-		--admin_password=$WP_ADMIN_PASSWORD \
-		--admin_email=$WP_ADMIN_MAIL \
-		--url=$WP_URL \
-		--allow-root
+if [ ! -f "var/www/wordpress/wp-config.php" ]
+then
+	wp core download --path=/var/www/wordpress --allow-root
+	
+	cd /var/www/wordpress
+	
+	wp config create --dbname=$SQL_DATABASE \
+			 --dbuser=$SQL_USER \
+			 --dbpass=$SQL_PASSWORD \
+			 --allow-root \
+			 --dbhost=mariadb:3306 --path='/var/www/wordpress'
+	
+	wp core install	--title=$WP_TITLE \
+			--admin_user=$ADMIN \
+			--admin_password=$WP_ADMIN_PASSWORD \
+			--admin_email=$WP_ADMIN_MAIL \
+			--url=$WP_URL --allow-root --skip-email
+	
+	wp user create $WP_USER $WP_USER_MAIL --role=author --user_pass=$WP_USER_PASSWORD --allow-root
+fi
 
-wp user create $WP_USER $WP_USER_MAIL --rol=author --user_pass=$WP_USER_PASSWORD --allow-root
+if [ ! -d "/run/php" ]
+then
+	mkdir /run/php
+fi
 
 cd -
-
-php-fpm7.4 -F
+	
+exec php-fpm7.4 -F
